refactor(interop): keep cruiser turbo reads in adapter
